feat(nextdns): add support for nextdns as resolver
Taking this a step further, it was nice to use Cloudflare and
`cloudflared` as my DNS resolver thanks to DNS over HTTPS, but t's
not all that configurable. This change brings in NextDNS support as
the DNS provider and configured it so it runs in another container, with
an override of the TTL by creating a new copy of `/var/nextdns/run.sh`.
Eventually that TTL should be configurable through and environment
variable and I should submit a patch to the upstream, but this works
now.

It should automatically fail back to Cloudflare when NextDNS is not
avialable.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>