feat(unifi): add support for Unifi Dream Machine Pro
This isn't really so much of a new feature as it is unbreaking an
existing set of features. The UDM Pro has some seriously broke DNS stuff
where it doesn't name hosts properly. This creates ANOTHER new container
that works for the local domain.

I still need to better document how to use this to get around some of
the reverse DNS issues when looking up in-addr.arpa PTRs, but it works
for at least the forward DNS.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>